Fix images with attributes not being selected
As of 3b6cf812, if `eileen` is defined as an object with the
`_choose_image_attributes()` method, and `eileen happy` is explicitly
defined as another normal image, and the author writes `show eileen
happy`, `eileen happy` will never be properly selected.

This commit fixes this by treating any image explicitly defined with one
or more attributes as it did before 3b6cf812, returning it on an exact
match in `apply_attributes()`. However, if the image is defined with no
attributes, and additionally has the `_choose_image_attributes()`
method, it will not be treated as an exact match as intended.